refator: project organization
